Updates to pokemonAnalysis and file structure
