+++ results/DecisionTree_important_features.csv
@@ -1,10 +1,14 @@
 ,Feature,Importance
-0,Total,0.706305949242483
-1,HP,0.09999958270634961
-4,Sp. Atk,0.06018050099682751
-2,Attack,0.04478509352458934
-5,Sp. Def,0.014652014652014654
-30,Type 2_Fighting,0.014453942111961833
-17,Type 1_Ground,0.013954299668585383
-23,Type 1_Steel,0.013954299668585383
-3,Defense,0.013289809208176557
+3,Sp. Atk,0.34035654987061414
+1,Attack,0.1714065445682721
+5,Speed,0.13543563497615885
+0,HP,0.1345491963094649
+2,Defense,0.04413265024978
+20,Type 1_Psychic,0.03725051522532155
+4,Sp. Def,0.02587323102837075
+41,Type 2_Water,0.020819098052613606
+28,Type 2_Fairy,0.019557334534273375
+23,Type 1_Water,0.019133441827601867
+33,Type 2_Grass,0.017766767411344607
+31,Type 2_Flying,0.016118011395571823
+35,Type 2_Ice,0.016118011395571823
