+++ run_all.sh
@@ -1,3 +1,3 @@
-python src/pokemonData.py "./data/Pokemon.csv" "./data/Pokemon_clean.csv"
-Rscript src/pokemonEDA.R ./data/Pokemon_clean.csv ./results
-python src/pokemonAnalysis.py "./data/Pokemon_clean.csv" "./results/DecisionTree_important_features"
+python src/1_pokemonData.py "./data/Pokemon.csv" "./data/Pokemon_clean.csv"
+Rscript src/2_pokemonEDA.R ./data/Pokemon_clean.csv ./results
+python src/3_pokemonAnalysis.py "./data/Pokemon_clean.csv" "./results/DecisionTree_important_features"
